[Ide] Run template actions when adding project to existing solution.
Fixed bug 26733 - Default set file in <Actions> is not getting open by
default when creating project in Workspace.
https://bugzilla.xamarin.com/show_bug.cgi?id=26733

Previously the project template actions would only be run when a new
solution was being created that was not being added to a workspace.

Now the project template actions are run when a new project is added
to an existing solution and if a new solution is added to an existing
workspace.