[NuGet] Detect project target framework change on reload.
Fixed bug 26710 - Migration tool does not update NuGet package
references
https://bugzilla.xamarin.com/show_bug.cgi?id=26710

The NuGet addin will not automatically update the NuGet package
references when the iOS migration tool migrates a project from
iOS Classic to iOS Unified, but it will detect the project's
target framework has changed when the project is reloaded.
The migration tool directly modifies the project file on disk
and then the project file will be reloaded into the IDE. When
it is reloaded the NuGet addin will check to see if the
project's target framework has changed and if it has then it
will check the compatibility of the NuGet packages and report
the results in the Package Console. NuGet packages that are
incompatible or need retargeting will be reported and a Retarget
menu option will be enabled on the Packages folder and on the
individual NuGet packages that can be retargeted. This Retarget
menu will uninstall and install the NuGet package so the correct
assemblies are referenced.