Syntax highlighting: Improve multiline comments Fixes #953.
It seems that when `// something *)` is found, that excludes the `*)` from being processed by multiline regexes.
The solution here is to match single line comments *until* but excluding '*)'.
This means '*)' is still available to the highlighter, but it also means that if you write '// comment *)' without having a starting '(*', everything after and including the '*)' will not get highlighted correctly.