Fix project to include roslyn assemblies.
**Note** that there are the following assemblies:
$(MSBuildExtensionsPath32)\..\4.5\Facades\System.Threading.Tasks.dll
$(MSBuildExtensionsPath32)\..\4.5\Facades\System.GLobalization.dll
$(MSBuildExtensionsPath32)\..\4.5\Facades\System.IO.dll
$(MSBuildExtensionsPath32)\..\4.5\Facades\System.Runtime.dll

This is because the interactive checker is not currently pcl aware, as
roslyn is a pcl assembly the implicit references needed by roslyn need
to be added here.  If we don’t then you get all sorts of errors in the
IDE, although compilation is fine as msbuild *is pcl aware.