[Ide] Fix New Project dialog browse button truncated on Linux.
Fixed bug 27652 - New "New solution" dialog has a button "Browse..."
which is cut off
https://bugzilla.xamarin.com/show_bug.cgi?id=27652

On Linux the font is larger and the text for the create gitignore
label was causing the Browse button to be truncated. Now the
labels extend into the last table column which contains the
Browse button so the button is no longer truncated.