[Designer.iOS] Make the keychain tests more robust
This will ensure we delete any pre-existing keychains before executing
the test suite.

This *should* be safe. The problem is that KeyChainOpen returns success
and also a 'valid' handle when a non-existent keychain is opened. This
means we have to wrap it in a try/catch as we never know if the Delete
call will throw or not.

Should fix a random failure on wrench.