[NUnit] Fix Go to failure menu always enabled in Unit Tests pad.
Fixed bug 28005 - 'Go to Failure' option is not working after right
click on Test project and tests.cs file
https://bugzilla.xamarin.com/show_bug.cgi?id=28005

Now the Go to Failure menu is only enabled if a test is selected and
it has failed. Previously the menu would always be enabled for any
tree node in the Unit Tests pad.