Fix hang in project builder
Changed the way we read the object reference from the output
of the build process. We need to keep reading the standard error.
The process will hang if something is written in that stream and nobody
reads it.

Fixes bug #27524 - Rebuilding app for different deployment target never
completes, Xamarin Studio hangs if you try to stop the build or quit