[NUnit] Update VB.NET project template icon name.
The name for the overlay icon was simplified in commit
9604482634c8dcd2b80421e25c2ee2a6c7663f73 but the VB.NET project
template was added on the new-project-dialog feature branch so
this change was not done to the VB.NET project template.

The VB.NET project template was using the old name, which still works,
but now it is consistent with the C# project template.