[NuGet] Keep Local Copy setting when updating a NuGet package.
Previously when updating a NuGet package the references would be
added with Local Copy set to true and the previous value would be
lost. MonoDevelop will generally have Local Copy set to false for
assemblies that are shared across MonoDevelop and since more NuGet
packages are being used in MonoDevelop this makes maintaining the
Local Copy value a lot more work.

Now if a reference is removed and then added during a NuGet update
the Local Copy value will be kept.

Note that currently this only works when updating the NuGet package
from the Solution window. If you add a NuGet package from the Add
Packages dialog then Local Copy is not preserved. This is a problem
for pre-release NuGet packages since you cannot update to a
new version of a pre-release NuGet package from the solution window.