[NuGet] Keep Local Copy setting when installing a NuGet package.
Local copy is now preserved if the Add Packages dialog is used to
install a different version of the NuGet package which is already
installed in the project. This allows pre-release NuGet packages to
keep the Local Copy setting since it is currently not possible to
update to a newer pre-release NuGet package from the Solution window.