Project execution improvements
Added PrepareExecution virtual to IBuildTarget. This method can be optionally
called before calling Execute, so that the project has a chance to start getting
everything ready for the execution while other operations (such as building the
project) are being done.

ProjectOperations.CheckAndBuildForExecute is now private. The Execute
methods in that class now call CheckAndBuildForExecute by default,
so it is not necessary to explicitly call it anymore.