[NuGet] Reduce status bar text length
Fixed bug #28730 - Status bar text from NuGet addin is too long
to fix in the status bar
https://bugzilla.xamarin.com/show_bug.cgi?id=28730

The new native status bar for Mac has a smaller width and the
status bar text from the NuGet addin would be too long to be
displayed. To reduce the length of the status bar text the
'Please see the Package Console for details.' text has been
removed.