Fix msbuild loading issue
Add support for loading msbuild files for projects that don't support the
msbuild engine. For those projects we can't use msbuild to evaluate
the project because evaluation is likely to fail due to missing targets.
In that case we use a fake msbuild engine which does a simple
evaluation (the same we did in the old project model).