[build] Don't forcibly link against the MacOS 10.8 SDK anymore
We originally did this because pre-release versions of Yosemite
completely broke scrolling in Xamarin Studio when we used the
MacOS 10.9 SDK to compile our stub launcher.

These issues seem to have gone away with current Yosemite versions
so we are going to lift the requirement to use the MacOS 10.8 SDK.

In testing this bug i compiled and tested with the following combos:

Compile on OSX 10.9.5, use the OSX10.9 SDK, run binary on OSX 10.9.
Compile on OSX 10.9.5, use the OSX10.9 SDK, run binary on OSX 10.10

Compile on OSX 10.10.2, use the OSX10.9 SDK, run binary on OSX 10.9.
Compile on OSX 10.10.2, use the OSX10.9 SDK, run binary on OSX 10.10

Compile on OSX 10.10.2, use the OSX10.10 SDK, run binary on OSX 10.9.
Compile on OSX 10.10.2, use the OSX10.10 SDK, run binary on OSX 10.10

All of the above worked perfectly.