[Core] Project model fixes
Don't automatically mark projects and solutions as shared after loading or
creating. Instead, SetShared() has to be called explicitly. The Ide calls
SetShared() on solutions opened in the workspace, so the result is more
or less the same.

Also added support for programatically attaching/detaching custom
extensions on projects and solutions.