[AspNet] Fix member completion as you type in Razor file
Typing in a variable, then dot, then some characters of the
member name, and then completing that member would insert the
characters typed in at the end of the completed text. The
completion list window would not show the matched member text
as highlighted either.

The problem was that the completion list was using an out of
date version of the hidden text editor used with Razor files.