Add support binding tasks to workspace objects
The new WorkspaceObject.BindTask() method allows binding a task to an
object. The object will track the execution of the task, and in case
the object is disposed, it will wait for the task to end. This method
also provides a cancellation token that is signalled when the object
is disposed, so that the task has a chance to end.