[Ide] Do not show Delete button when removing missing file.
Fixed bug #26409 - Removing file from Solution Pad
https://bugzilla.xamarin.com/show_bug.cgi?id=26409

When removing a file from Solution Pad when the file does not exist
the Delete button is now no longer displayed in the confirmation
dialog. Since the file cannot be deleted and can only be removed from
the project the dialog will show only the Cancel and Remove from
Project buttons. Also the dialog will no longer display the secondary
message about the difference between the Delete and Remove from
Project button if the file does not exist.

If the file being removed has a dependent child file then the Delete
button will not be displayed if the child and its parent do not exist.

When removing a file link the Delete button is not shown since the
existing code does not delete the original file but only removes the
link from the project.