[AspNet] Fix Razor completion not working when project opened.
If a Razor file (.cshtml) was left open in a project and then the
IDE is closed, re-opened and the project opened again, the
code completion for the Razor would not work until the file is
either closed and re-opened or another Razor file was opened.

This problem occurred because the RazorCSharpParser was not
thread safe and stores parser information as member variables.
Multiple parser threads try to parse the document at the same time
causing one thread to wait for an auto reset event to fire whilst
the other finished. Then on typing into the editor a new thread
would start a parse and be blocked by the auto reset event whilst
the previously held thread would complete.

The RazorCSharpParser has been changed so only one thread can
parse a particular document at a time.