[Ide] Avoid crash trying to reload a project
The MD workspace subscribes the project modified events just before the
project is added to the workspace. It may happen that the project is modified
after the modified event is subscribed, but before the project is added to
the workspace. In that case the workspace will try to reload the project and
it will crash because the project doesn't yet exist in the solution.

As a workaround, the workspace now checks if the project belongs to the
workspace before trying to reload it.