[Git] Saner stashing handling.
This is all sorts of bonkers. We need a way to differentiate between a
null stash that was caused by a credentials abort and a null stash that
was caused by no changes in the index/workdir.

Therefore, TryCreateStash (probably TryStash) in the future. true/false
for abort, null stash for abort/empty.

Also, fixed the SwitchToBranch handler to use PopStash not manual merging
of the index.