[NUnit] Fix UITests failing when run from the Unit Tests window
The NUnitService's TestSessionStarted event contained a TestSession
with a Task that returned immediately and not the actual Task
that would be run. This has been changed so when the
TestSessionStarted event fires the TestSession's Task is the
task that will be started later so it can be used to be notified
when the Task has completed.