[UITest] Save Ide.log for failed tests and delete screenshots for passed tests
As Alan suggested, it would be wise to only deal with data related
to failed tests. So in this change, we check if the test has passed,
if passed then we delete the screenshot foldet for that test.

Before XS starts, we set the Log path and level, then in test teardown
we check if the test passed, if passed then we delete the log

In the FixtureTearDown we delete empty directories such that after
the test finishes, we are left with screenshots and logs of only
failed tests