[Ide] Fix null reference when creating new F# file.
Partial fix for bug #30365 - Error when adding new file in F# project
https://bugzilla.xamarin.com/show_bug.cgi?id=30365

This fixes the null reference when adding a new empty F# class file
from the template. However the F# formatter is currently returning
an empty string since the F# code generated by the F# code dom
is incomplete and the formatter cannot format it. This means the
class file generated is empty. To fix this we can either change the
F# formatter to return the original text if it cannot format it
or we change the AbstractCodeFormatter to allow a null to be
returned from the FormatImplementation method which has been changed
from returning a string to returning an ITextSource.