[Ide] Do not use code dom file templates for F# projects.
The F# code dom generates incomplete F# code for the empty class,
struct, interface and enum file templates which the F# code
formatter cannot format so the files are created as empty text. To
avoid this these code dom file templates are not displayed for F#
projects in the New File dialog.