[AutoTest] When a timeout occours, raise TimeoutException instead of generic Exception
Raising TimeoutException in the case of timeout can help us narrow down the
causes of failures because such exceptions can be filtered and analyzed
seperately. Lots of timeout exception happen because the test machine
is slower or the network is slower which leads to test failures even when
there is nothing wrong with the tests.