[AutoTest] Added SourceQuery to QueryResult so that actual query can be known during timeout
Right now we can catch the exception and know the exact query only in
method ExecuteQuery, WaitForElement and WaitForNoElement because
in those methods we are provided with original AppQuery.

In methods like Select, Click, EnterText and Toggle, we just get
the AppResult and hence it becomes hard to figure out the exact
query which led to this result.

In this commit, SourceQuery is a string which holds the original
string query which was assigned to it during evaluation. This
is an additional data useful for debugging when there is a
timeout in the above four methods mentioned