[NuGet] Packages folder not updating after package removed.
Fixed bug #30599 - Removing a nuget package does not refresh the
packages list
https://bugzilla.xamarin.com/show_bug.cgi?id=30599

With two or more solutions opened at the same time the Packages folder
would not be updated for all solutions when a NuGet package was
removed. Only projects with one solution would have their Packages
folder in the Solution window updated. The projects in the other
solutions would not update after a NuGet package was removed. This is
because the NuGet addin was not refreshing the Packages folder node
for all solutions currently open.