[NuGet] Fix packages.config marked as deleted by Git on package update
Fixed bug #27035 - After updating Xamarin.Forms package the
packages.config file show red icon.
https://bugzilla.xamarin.com/show_bug.cgi?id=27035

The simplest way to reproduce the original problem is to have an old
pre-release version of a NuGet package installed (e.g. Xamarin.Forms
1.4.3.6356-pre1), note that this must be the only NuGet package in the
project, make sure everything is committed and update to date
in Git, then open the Add Packages dialog again and find a newer
pre-release version (e.g. Xamarin.Forms 1.4.3.6358-pre2) and install
that. In previous versions of the NuGet addin the packages.config
file would end up being shown by Git as having been deleted. Now with
this commit the packages.cpnfig file is shown as being updated.

The problem was that installing a newer version of the NuGet package,
when the packages.config file only had one NuGet package, the
packages.config file will be deleted as the old version of the NuGet
package is uninstalled and re-created when the new version is
installed. The file being deleted was being reported to the version
control system causing it to be marked as deleted by Git. Now the
file deletion is not reported to the version control system since it
will be re-created by NuGet.