[NuGet] Fix packages.config marked as deleted by Git on retargeting
Similar to the bug fixed in 1fd0463819aedab16bf36b578091a26e652dc04f

With a project that had only one NuGet package in its packages.config
file, when the project's target framework was changed, the Retarget
menu option that is made available will uninstall and then install the
package so the correct assemblies are referenced for the project's
target framework. The packages.config file is deleted since all the
NuGet packages are removed from the project. This was causing Git to
mark the file as deleted. Now the packages.config file is deleted by
the NuGet addin and the version control system is not told about this
since the file will be recreated when the NuGet package is installed
again. This prevents Git showing the file as deleted.