[Core] Fix proxy credentials not requested for https url.
When a system wide proxy is configured for all https requests then
a https url would fail to request proxy credentials on Mono and the
user will not be prompted. Instead the request would be retried 10
times and then fail.

This only affects https requests. Requests to http url will prompt
for proxy credentials.

The problem is a difference in behaviour of Mono compared with
Microsoft's .NET Framework when a proxy requires authentication and
a https url is requested:

https://bugzilla.xamarin.com/show_bug.cgi?id=19594

When connecting to a https url through a proxy that requires
authentication when using Microsoft's .NET Framework a WebException
is thrown with:

    WebException.Status = ProtocolError
    WebException.Response.StatusCode = ProxyAuthenticationRequired

On Mono a WebException is thrown with:

    WebException.Status = SecureChannelFailure
    WebException.Response = null

So now a check is made when not running on Windows to see if the
exception is a proxy authentication error.