[Ide] Prevent solution being added twice to same workspace.
Adding the same solution to a workspace was failing with an error:

Save failed. Collection was modified; enumeration operation may not execute.
System.InvalidOperationException: Collection was modified; enumeration operation may not execute.

Now a check is made that the solution or workspace does not already
exist in the workspace. If it does exist then a message is shown and
the solution/workspace is not added.