[NuGet] Do not clear update info if NuGet package is installed.
If a NuGet package is installed whilst check for updates is being
done then the update information was being cleared in the Solution
window. So when the check for updates completed it would not show an
update for the package that finished installing just after the check
started.

Now the update information is only cleared for a NuGet package if
it has been uninstalled. This ensures the count is correct on the
Packages folder but still allows an update to be shown for a NuGet
package that was installed or updated whilst the check for updates
was running.

This is to improve the problem reported in bug 30967 where a
NuGet package which is being installed as the check for updates
occurred, but was in the packages.config file before the update
started, would not show any update information in the Solution
window after check for updates had finished.

https://bugzilla.xamarin.com/show_bug.cgi?id=30967

It also means that if a newer NuGet package is installed but not the
latest then the update information will still be displayed in the
Solution window since there is a newer update available.