[Ide] Handle recent project paths containing url encoded string.
Fixed bug #30252 - Not able to open recent project if solution path
contains "%20"
https://bugzilla.xamarin.com/show_bug.cgi?id=30252

If the path to a solution contained %20 then it would not be possible
to open the solution from the welcome page. Instead a file not found
message would be displayed with the %20 characters replaced by spaces.

The problem was that the solution filename saved in the recent
projects list was not url encoded but it was loaded into a Uri
before trying to open the solution from the recent projects list
on the main welcome page. Now the filename is not loaded into a Uri
so the filename is not decoded.