[NuGet] Stop solution being closed when NuGet packages are installing.
Fixed bug #26243 - Possible to close a solution whilst NuGet packages
are added and the solution is not fully created
https://bugzilla.xamarin.com/show_bug.cgi?id=26243

Previously when a solution was closed when a NuGet package is being
installed then installation would continue and you would see error
messages saying "Reference to unknown project 'test' ignored". This
can happen when creating a new project that has several NuGet packages
such as the ASP.NET MVC with Tests project template.

Now if the user tries to close the solution or quit the application
when a NuGet package is being installed then a message will appear
saying that it is not possible to close the solution whilst NuGet
packages are being processed. The message can appear when closing
the solution if a NuGet package is being updated, installed or
removed. Basically if any background NuGet task is being run then
the solution cannot be closed. Checking for package updates can be
cancelled and will not prevent the solution from being closed.