[MSBuild] Fix performance summary shown for non-diagnostic logging.
Fixed bug #31189 - Windows - Diagnostic log is displayed after user
changes Log verbosity to from Diagnostic to Quiet.
https://bugzilla.xamarin.com/show_bug.cgi?id=31189

On Windows when the Build log verbosity is set to Diagnostic then
MSBuild switches on the performance summary which can be seen in the
build output. Switching back to Quiet you still see the performance
summary information. This is only a problem on Windows.

The problem is that the ConsoleLogger is re-used on each build and
once the log verbosity is set to Diagnostic the performance summary
is always output from MSBuild on Windows. The problem can be fixed by
closing and re-opening the solution or by killing the
MonoDevelop.Projects.Formats.MSBuild.exe process. There does not seem
to be a way to get the MSBuild ConsoleLogger to stop showing the
performance summary by updating one of its properties so it is now
recreated on each build which resets the state and prevents the
performance summary information being displayed unless the log
verbosity is set to Diagnostic.