[Ide] Prevent the same solution being added to an existing workspace.
Fixed bug #20639 - InvalidOperationException displayed on adding
solution with same name.
https://bugzilla.xamarin.com/show_bug.cgi?id=20639

Steps to reproduce:
1. Open or create a Workspace
2. Add new solution (empty or non empty) say 'Sol1'.
3. Add another solution with same name 'Sol1'

Result:
InvalidOperationException: Collection was modified; enumeration
operation may not execute.

Saving the solution that was already in the workspace caused the
solution to be re-loaded and the workspace items collection was
modified whilst the workspace was iterating over each of its items
to save them. To prevent this the New Project dialog now checks the
parent workspace to see if the solution filename already exists and
prevents the solution from being created. This is similar to the check
that is made when adding a new project to an existing solution.