[Windows] Fix ArgumentOutOfRangeException in Choose encodings window.
Fixed bug #31402 - Crash when click 'Drown/Up' button multiple times
in "Choose encodings" window.
https://bugzilla.xamarin.com/show_bug.cgi?id=31402

To repro:
1. Click on File >> Open option.
2. Select "Add or Remove" option from Encoding field.
3. Open Choose encoding window.
4. Select any item not at the end of the "Encoding shown in menu" list
5. Click on "Drown/Up" button multiple times.

Then an ArgumentOutOfRangeException was thrown.

The problem was that the up/down buttons in the Choose encodings
window were not being disabled after they were clicked when the item
currently selected had moved to the top or bottom of the list.