[NuGet] Support multiple packages.config files in same directory.
NuGet supports multiple packages.config files in the same directory
if they are each named after their corresponding project in the
directory. The file name is of the form:

    packages.ProjectName.config

This commit provides support for finding a packages.config file named
after the project. Previously the solution window would not show the
project as having any NuGet packages. Also the package restore would
not restore any packages.