[Mac] Force parent window on to native select file dialogs
This caused a infinite loop inside cocoa which is caused by not
having a decent window chain. It probably took in the main window
as the window to be hidden but the child dialog of the main window
was still being shown, therefore cocoa blew up.

Bug 31448 - [Shell] XS crashes when Cmd+H is pressed in the Open File
Dialog from Edit References > Select Assembly