[Test Framework] Remove the IsBuildSuccessful calls from AutoTestSession
IsBuildSuccessful is not right for being in AutoTestSession as it is an
interpretation of the data, which should be left up to the tests to decide. It
has been replaced with a method called ErrorCount which returns the number of
errors reported at a specified severity. It is then up to the caller to decide
what to do with this data.