[Mac] Hide Close workspace check box when solution viewer not selected
Fixed bug #31489 - [Windows] Close workspace check box is hidden
when Source Code Editor/Hex Editor is selected in File to Open dialog
https://bugzilla.xamarin.com/show_bug.cgi?id=31489

The bug description is incorrect. The problem is that the Windows and
Mac File to Open dialog are inconsistent. The Mac open file dialog has
been changed so the Close current workspace check box is only
displayed if the currently selected viewer (Open with) is
"Solution Workbench". If another viewer is selected, for example the
Source Code Editor, then a project file or solution file can be
opened into the text editor without the existing solution open in the
workspace needing to be closed. This makes the behaviour consistent
with how it works on Windows.