[NuGet] Fix null ref exception when solution closed.
Fixes bug #21845 - NRE in MonoDevelop.PackageManagent when solution
closed
https://bugzilla.xamarin.com/show_bug.cgi?id=21845

Currently the NuGet addin does not handle multiple solutions open in
the workspace. On closing the workspace with two or more solutions
a null reference exception would occur when the second solution was
unloaded. This is now handled but the logic needs to be improved since
currently with two solutions being open in the workspace only one
solution will show NuGet package compatibility problems when the
target framework of one of its projects is modified.