[Windows] Disable encoding if solution selected in Open File dialog.
Fixed bug #31612 - Mac - 'Auto Detected' dropdown is disabled after
selecting any .sln file at "File to Open" window.
https://bugzilla.xamarin.com/show_bug.cgi?id=31612

The bug description is incorrect. The problem is that on Windows the
File to Open dialog was not disabling the Encoding combo box when the
Solution Workbench is selected to open the file with. On the Mac the
encoding combo box was disabled. The encoding is ignored if the file
is a solution or project file and is opened when Solution Workbench is
selected. So the Windows File to Open dialog has been changed to
disable the encoding combo box when Solution workbench is selected to
make it consistent with the behaviour on the Mac.