[Ide] Do not add Solution Items folder when add files is cancelled.
It was possible to cancel adding a file as a solution item but the
Solution Items folder would still be added to the solution file.

The steps to reproduce this:

1. Right clicking the solution.
2. Select Add - Add Files...
3. Browse to a file that is outside the solution's directory and
select Open.
4. Select Cancel on the dialog which is then displayed saying that the
file is outside the folder directory and asks whether the file should
be copied, moved or linked to.

A Solution Items folder would be added to the solution even though the
operation was cancelled.