Don't use extension methods for Theming class
In F# (and maybe C# too), assemblies of types used in extension methods must
be explicitly referenced. Since the ApplyTheme method was using NSWindow
as parameter, the F# addin failed because it wasn't referencing Xammac.

To avoid this reference (which is not always possbile), the extension methods
have been converted to regular static methods.