Improve selection of project to attach to a document
Improved the logic that selects the project to be attached to a document,
when the file is included in several projects (either through a link or through
a shared project). The changes are:
1) When looking for the default projects, ignore projects that are disabled
for the current configuration.
2) If the current configuration changes and the project attached to the
document is now disabled, assign another project.
3) If the current startup project of the solution changes, and the new
startup project includes the file, then set it as current project.

Fixes bug #32037 - Navigation bar's project selector defaults to first
project in solution.