[NuGet] Do not show status bar warnings about PowerShell scripts.
Fixed bug #32577 - Do not show warning in status bar about PowerShell
scripts when installing/updating a NuGet package
https://bugzilla.xamarin.com/show_bug.cgi?id=32577

Previously if a NuGet package had a PowerShell script the status bar
would indicate that the NuGet package had installed with warnings.
Now the warning is not displayed in the status bar. A warning will
still be logged in the Package Console.