[Ide] Select last project template used in New Project dialog.
Fixed bug #32637 - Remember the last selected section and sub-section
in new project dialog
https://bugzilla.xamarin.com/show_bug.cgi?id=32637

Previously the top two categories were saved and remembered between
sessions, but the last used template was not and the first template
in the second category would be selected.

Now the last project template used is selected in the New Project
dialog. This information is saved between sessions so when the
IDE is closed and re-opened the last project template used will be
selected again in the New Project dialog.