[Ide] Ensure last selected template is on screen in New Project dialog
In the Other/Misc category all the project templates do not fit
on screen. If the last selected project template was off screen then
on opening the New Project dialog again the template was still off
screen but selected. Now the New Project dialog ensures that the
last selected project template is on screen by scrolling if necessary.