[NuGet] Prefix PowerShell script message with WARNING text.
Since a warning is no longer shown in the status bar when installing a
NuGet package that has PowerShell scripts the message in the Package
Console is now prefixed with WARNING: to make it easier to find.

A better improvement would be to colour the text or its background so
it stands out from the normal text but that has not been implemented.