[Ide] Handle groups for previously selected project template in dialog
Check all project templates in a group when looking for the template
previously selected in the New Project dialog. The parent can be
different if the addins are loaded in a different order. Previously
only the parent template was checked for an id match which worked but
now the code is more defensive and will handle a different parent
template if the templates are grouped.